+++ app/Http/Controllers/Farming/GuarantorController.php
@@ -22,9 +22,13 @@ class GuarantorController extends Controller
      */
     public function index()
     {
-        $guarantors = Guarantor::where('created_by', Auth::user()->id)->get();
+        if (\Auth::user()->can('manage-farmer_guarantor')) {
+            $guarantors = Guarantor::where('created_by', Auth::user()->id)->get();
 
-        return view('admin.farmer.guarantor.index', compact('guarantors'));
+            return view('admin.farmer.guarantor.index', compact('guarantors'));
+        } else {
+            return redirect()->back()->with('error', 'Permission denied.');
+        }
     }
 
     /**
@@ -32,14 +36,18 @@ class GuarantorController extends Controller
      */
     public function create()
     {
-        $farmings = Farming::query()->select('farmings.*')->join('users', 'users.id', 'farmings.created_by')
-            ->where('farmings.is_validate', 1)
-            ->where('farmings.created_by', Auth::user()->id)
-            ->orWhere('users.supervisor_id', Auth::user()->id)
-            ->get();
+        if (\Auth::user()->can('create-farmer_guarantor')) {
+            $farmings = Farming::query()->select('farmings.*')->join('users', 'users.id', 'farmings.created_by')
+                ->where('farmings.is_validate', 1)
+                ->where('farmings.created_by', Auth::user()->id)
+                ->orWhere('users.supervisor_id', Auth::user()->id)
+                ->get();
 
-        $countries = Country::all();
-        return view('admin.farmer.guarantor.create', compact('countries', 'farmings'));
+            $countries = Country::all();
+            return view('admin.farmer.guarantor.create', compact('countries', 'farmings'));
+        } else {
+            return redirect()->back()->with('error', 'Permission denied.');
+        }
     }
 
     /**
@@ -47,27 +55,30 @@ class GuarantorController extends Controller
      */
     public function store(Request $request)
     {
-
-        try {
-            $this->validate($request, [
-                'farming_id' => 'required',
-                'name' => 'required',
-                'father_name' => 'required',
-                'country_id' => 'required',
-                'state_id' => 'required',
-                'district_id' => 'required',
-                'block_id' => 'required',
-                'gram_panchyat_id' => 'required',
-                'village_id' => 'required',
-                'post_office' => 'required',
-                'police_station' => 'required',
-                'age' => 'required',
-                'created_by' => 'required',
-            ]);
-            Guarantor::create($request->all());
-            return redirect()->to(route('admin.farmer.guarantor.index'))->with('success', 'Guarantor Added Successfully.');
-        } catch (Exception $e) {
-            return redirect()->back()->with('error', $e->getMessage());
+        if (\Auth::user()->can('create-farmer_guarantor')) {
+            try {
+                $this->validate($request, [
+                    'farming_id' => 'required',
+                    'name' => 'required',
+                    'father_name' => 'required',
+                    'country_id' => 'required',
+                    'state_id' => 'required',
+                    'district_id' => 'required',
+                    'block_id' => 'required',
+                    'gram_panchyat_id' => 'required',
+                    'village_id' => 'required',
+                    'post_office' => 'required',
+                    'police_station' => 'required',
+                    'age' => 'required',
+                    'created_by' => 'required',
+                ]);
+                Guarantor::create($request->all());
+                return redirect()->to(route('admin.farmer.guarantor.index'))->with('success', 'Guarantor Added Successfully.');
+            } catch (Exception $e) {
+                return redirect()->back()->with('error', $e->getMessage());
+            }
+        } else {
+            return redirect()->back()->with('error', 'Permission denied.');
         }
     }
 
@@ -84,28 +95,32 @@ class GuarantorController extends Controller
      */
     public function edit($id)
     {
-        $guarantor = Guarantor::find($id);
-        $countries = Country::all();
-        $states = State::where('country_id', $guarantor->country_id)->get();
-        $districts = District::where('state_id', $guarantor->state_id)->get();
-        $blocks = Block::where('district_id', $guarantor->district_id)->get();
-        $gram_panchyats = GramPanchyat::where('block_id', $guarantor->block_id)->get();
-        $villages = Village::where('gram_panchyat_id', $guarantor->gram_panchyat_id)->get();
-        $farmings = Farming::query()->select('farmings.*')->join('users', 'users.id', 'farmings.created_by')
-            ->where('farmings.is_validate', 1)
-            ->where('farmings.created_by', Auth::user()->id)
-            ->orWhere('users.supervisor_id', Auth::user()->id)
-            ->get();
-        return view('admin.farmer.guarantor.edit', compact(
-            'guarantor',
-            'countries',
-            'states',
-            'districts',
-            'blocks',
-            'gram_panchyats',
-            'villages',
-            'farmings',
-        ));
+        if (\Auth::user()->can('edit-farmer_guarantor')) {
+            $guarantor = Guarantor::find($id);
+            $countries = Country::all();
+            $states = State::where('country_id', $guarantor->country_id)->get();
+            $districts = District::where('state_id', $guarantor->state_id)->get();
+            $blocks = Block::where('district_id', $guarantor->district_id)->get();
+            $gram_panchyats = GramPanchyat::where('block_id', $guarantor->block_id)->get();
+            $villages = Village::where('gram_panchyat_id', $guarantor->gram_panchyat_id)->get();
+            $farmings = Farming::query()->select('farmings.*')->join('users', 'users.id', 'farmings.created_by')
+                ->where('farmings.is_validate', 1)
+                ->where('farmings.created_by', Auth::user()->id)
+                ->orWhere('users.supervisor_id', Auth::user()->id)
+                ->get();
+            return view('admin.farmer.guarantor.edit', compact(
+                'guarantor',
+                'countries',
+                'states',
+                'districts',
+                'blocks',
+                'gram_panchyats',
+                'villages',
+                'farmings',
+            ));
+        } else {
+            return redirect()->back()->with('error', 'Permission denied.');
+        }
     }
 
     /**
@@ -113,9 +128,13 @@ class GuarantorController extends Controller
      */
     public function update(Request $request, $id)
     {
-        $guarantor = Guarantor::find($id);
-        $guarantor->update($request->all());
-        return redirect()->back()->with('success', 'Guarantor Updated Successfully.');
+        if (\Auth::user()->can('edit-farmer_guarantor')) {
+            $guarantor = Guarantor::find($id);
+            $guarantor->update($request->all());
+            return redirect()->back()->with('success', 'Guarantor Updated Successfully.');
+        } else {
+            return redirect()->back()->with('error', 'Permission denied.');
+        }
     }
 
     /**
@@ -123,8 +142,12 @@ class GuarantorController extends Controller
      */
     public function destroy($id)
     {
-        $guarantor = Guarantor::find($id);
-        $guarantor->delete();
-        return redirect()->back()->with('success', 'Guarantor Deleted Successfully.');
+        if (\Auth::user()->can('delete-farmer_guarantor')) {
+            $guarantor = Guarantor::find($id);
+            $guarantor->delete();
+            return redirect()->back()->with('success', 'Guarantor Deleted Successfully.');
+        } else {
+            return redirect()->back()->with('error', 'Permission denied.');
+        }
     }
 }
